project/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
sqlite3
bcrypt==4.0.1
jwt==1.3.1
PyJWT==2.8.0
qrcode==7.4.2
pillow==10.0.1
python-dotenv==1.0.0
werkzeug==2.3.7